+++ engine/__init__.py
@@ -1,1 +1,1 @@
-# Hermes Pet Win - Engine Package
+# BuddyDesk - Engine Package